requirements.txt
# Server MCP Inventario - Dipendenze per Claude Desktop
# MCP Framework
mcp[cli]>=1.2.0

# Database
asyncpg>=0.28.0

# Configuration
python-dotenv>=1.0.0

# HTTP Client (per future estensioni)
httpx>=0.25.0

# Logging e utilities
typing-extensions>=4.0.0
